# target/classes/api/api.yml
openapi: 3.0.1
info:
  title: Herramientas de números primos
  description: Microservicio que contiene algunas herramientas para visualizar o trabajar con números primos
  version: "1.0.0"

servers:
  - url: http://localhost:8080
    description: Servidor local

paths:
  /v0/polar:
    post:
      summary: Lanza una función de cálculo de las coordenadas polares de los primeros x primos.
      operationId: polarCoordenates
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CribaRequest'
      responses:
        '200':
          description: Cálculos realizados con éxito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PolarResponse'
        '400':
          description: Petición inválida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Demasiadas peticiones en poco tiempo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: API de cálculo no disponible o petición que consume más recursos que los disponibles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /v0/diff:
    post:
      summary: Lanza una función de cálculo de las diferencias entre los primeros x primos.
      operationId: primeDiff
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CribaRequest'
      responses:
        '200':
          description: Cálculos realizados con éxito.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DiffResponse'
        '400':
          description: Petición inválida.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '429':
          description: Demasiadas peticiones en poco tiempo.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '500':
          description: Error interno del servidor.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '503':
          description: API de cálculo no disponible o petición que consume más recursos que los disponibles.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

components:
  schemas:
    CribaRequest:
      type: object
      required:
        - number
      properties:
        number:
          type: number
          example: 10

    PolarResponse:
      type: array
      items:
        type: object
        required:
          - axisX
          - axisY
      properties:
        axisX:
          type: number
          example: 5
        axisY:
          type: number
          example: 7
    
    DiffResponse:
      type: object
      required:
        - primos
        - diferencias
        - mensaje
      properties:
        primos:
          type: array
          items:
            type: number
        diferencias:
          type: array
          items:
            type: number
        mensaje:
          type: string
    
    ErrorResponse:
      type: object
      required:
        - code
        - status
        - message
      properties:
        code:
          type: number
        status:
          type: string
        message:
          type: string
        timestamp:
          type: number